Use a named pipe for logging in upstart example
The ability to log directly to /dev/kmsg was removed in Ubuntu 14.04,
making the existing example config fail to perform any real logging.
Appending a pipe to the last `exec` within the `script` block gives
upstart an incorrect PID to work with, which can lead to various issues.

This uses a FIFO solution that logs to syslog at any permission level
and doesn't leave anything on the file system afterwards.