Create separate data structures for queue and worker
Rspec allows expectations to be written using change(LogWorker.jobs,
:size). The result of .jobs in this case was a derived array that wasn't
manipulated under the covers. So when Rspec went to check the size of
the array after the fact, it appeared that nothing had changed.

This sets up a true array for the jobs for a single worker and
pulls jobs off that array so rspec can properly make assertions.