Handle errors raised by the reloader (#3212)
The reloader might raise an error instead of yielding control. If this
happens, the job will silently fail to run, without being reenqueued and
without logging any indication of what happened.

This change doesn't prevent the job from being lost, but at least if the
error handlers are called it will be possible to diagnose the problem.